Fix : 스플래시 화면을 액티비티에서 프래그먼트로 변경 - 스플래시 액티비티 삭제 - 스플래시 프래그먼트 시작 시 꿀팁리스트와 오늘의 날씨를 가져오도록 구현 - 메인 액티비티부터 시작하도록 manifest 수정 - 꿀팁 리스트를 가져오는 메소드 HomeRepository -> MainRepository로 변경 - 오늘의 날씨를 가져오는 메소드 HomeRepository -> MainRepository로 변경 - HomeViewModel의 꿀팁 리스트를 가져오는 코드 MainViewModel로 변경 - HomeViewModel의 오늘의 날씨를 가져오는 코드 MainViewModel로 변경 - nav_graph.xml 의 시작지점을 homeFragment에서 splashFragment로 변경 #6
