+++ app/src/main/AndroidManifest.xml
@@ -15,18 +15,13 @@
         android:theme="@style/Theme.GumiLifeProject"
         tools:targetApi="31">
         <activity
-            android:name=".ui.splash.SplashActivity"
-            android:exported="true" >
+            android:name=".ui.main.MainActivity"
+            android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
-        </activity>
-        <activity
-            android:name=".ui.main.MainActivity"
-            android:exported="true">
-
         </activity>
     </application>
