+++ app/src/main/java/com/ssafy/gumi_life_project/data/model/Tip.kt
@@ -1,8 +1,13 @@
 package com.ssafy.gumi_life_project.data.model
 
+import com.google.gson.annotations.SerializedName
+
 data class Tip(
     val id : Int,
     val subject : String,
-    val discription : String,
+    @SerializedName("discription")
+    val description : String,
     val good : Int
-)
+) {
+    constructor() : this(0, "", "", 0)
+}
