+++ app/src/main/java/com/ssafy/gumi_life_project/ui/home/HomeFragment.kt
@@ -5,11 +5,13 @@ import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.fragment.app.activityViewModels
 import androidx.fragment.app.viewModels
 import com.ssafy.gumi_life_project.R
 import com.ssafy.gumi_life_project.data.model.Tip
 import com.ssafy.gumi_life_project.databinding.FragmentHomeBinding
 import com.ssafy.gumi_life_project.ui.home.crosswalk.CrossWorkBottomSheet
+import com.ssafy.gumi_life_project.ui.main.MainViewModel
 import com.ssafy.gumi_life_project.util.template.BaseFragment
 import dagger.hilt.android.AndroidEntryPoint
 import kotlin.random.Random
@@ -20,6 +22,7 @@ class HomeFragment : BaseFragment<FragmentHomeBinding>(
     R.layout.fragment_home
 ) {
     private val viewModel by viewModels<HomeViewModel>()
+    private val activityViewModel by activityViewModels<MainViewModel>()
 
     override fun onCreateBinding(
         inflater: LayoutInflater,
@@ -28,13 +31,8 @@ class HomeFragment : BaseFragment<FragmentHomeBinding>(
         return FragmentHomeBinding.inflate(inflater, container, false).apply {
             lifecycleOwner = viewLifecycleOwner
             viewModel = this@HomeFragment.viewModel
+            mainViewModel = this@HomeFragment.activityViewModel
         }
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-        viewModel.getAllTipList()
-        viewModel.getNowWeather()
     }
 
     override fun init() {
@@ -61,7 +59,9 @@ class HomeFragment : BaseFragment<FragmentHomeBinding>(
                     bottomSheetDialogFragment.show(childFragmentManager, "CrossWorkBottomSheet")
                 }
             }
+        }
 
+        with(activityViewModel) {
             tip.observe(viewLifecycleOwner) { event ->
                 event.getContentIfNotHandled()?.let {
                     val randomTip = getRandomTip(it)
@@ -74,7 +74,7 @@ class HomeFragment : BaseFragment<FragmentHomeBinding>(
 
                 }
             }
-            
+
             weather.observe(viewLifecycleOwner) { event ->
                 event.getContentIfNotHandled()?.let {
                     val weather = it.data
@@ -82,9 +82,8 @@ class HomeFragment : BaseFragment<FragmentHomeBinding>(
                     bindingNonNull.textviewTodayWeatherTemperature.text = weather.temperature + "º"
                     makeWeatherIcon(weather.precipitationType)
                 }
-                
-            }
 
+            }
         }
     }
