+++ app/src/main/java/com/ssafy/gumi_life_project/ui/main/MainViewModel.kt
@@ -1,13 +1,20 @@
 package com.ssafy.gumi_life_project.ui.main
 
+import android.util.Log
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.viewModelScope
 import com.ssafy.gumi_life_project.data.model.Event
+import com.ssafy.gumi_life_project.data.model.Tip
+import com.ssafy.gumi_life_project.data.model.WeatherResponse
 import com.ssafy.gumi_life_project.data.repository.main.MainRepository
+import com.ssafy.gumi_life_project.util.network.NetworkResponse
 import com.ssafy.gumi_life_project.util.template.BaseViewModel
 import dagger.hilt.android.lifecycle.HiltViewModel
+import kotlinx.coroutines.launch
 import javax.inject.Inject
 
+private const val TAG = "MainViewModel_구미"
 @HiltViewModel
 class MainViewModel @Inject constructor(
     private val repository: MainRepository
@@ -15,6 +22,62 @@ class MainViewModel @Inject constructor(
 
     private val _msg = MutableLiveData<Event<String>>()
     val errorMsg : LiveData<Event<String>> = _msg
+
+    private val _tip = MutableLiveData<Event<List<Tip>>>()
+    val tip: LiveData<Event<List<Tip>>> = _tip
+
+    private val _weather = MutableLiveData<Event<WeatherResponse>>()
+    val weather: LiveData<Event<WeatherResponse>> = _weather
+
+    fun getAllTipList() {
+        showProgress()
+        viewModelScope.launch {
+            val response = repository.getAllTipList()
+
+            val type = "정보 조회에"
+            when (response) {
+                is NetworkResponse.Success -> {
+                    _tip.value = Event(response.body)
+                }
+                is NetworkResponse.ApiError -> {
+                    postValueEvent(0, type)
+                }
+                is NetworkResponse.NetworkError -> {
+                    postValueEvent(1, type)
+                }
+                is NetworkResponse.UnknownError -> {
+                    postValueEvent(2, type)
+                }
+            }
+        }
+        hideProgress()
+    }
+
+    fun getNowWeather() {
+        showProgress()
+        viewModelScope.launch {
+            val response = repository.getNowWeather()
+            Log.d(TAG, "getNowWeather: $response")
+
+            val type = "정보 조회에"
+            when (response) {
+                is NetworkResponse.Success -> {
+                    _weather.value = Event(response.body)
+                    Log.d(TAG, "getNowWeather: ${weather.value}")
+                }
+                is NetworkResponse.ApiError -> {
+                    postValueEvent(0, type)
+                }
+                is NetworkResponse.NetworkError -> {
+                    postValueEvent(1, type)
+                }
+                is NetworkResponse.UnknownError -> {
+                    postValueEvent(2, type)
+                }
+            }
+        }
+        hideProgress()
+    }
 
 
     private fun postValueEvent(value : Int, type: String) {
